move get_query_arrows() from RelationshipSchema to Query
